Add makefiles, folders and test files
This is the actual template, containing folders and files. Both source
and make files. The whole project is built form the top directory. To
build the program in src/, simply run `make` or `make prog` in the top
directory. To build the test program in test/, run `make test`. Both
`make prog` and `make test` will create binaries in the bin/ directory.

The structure of the template is given below:
.
|-- .gitignore
|-- Makefile
|-- README
|-- include/
|   `-- implementation.h
|-- src/
|   |-- Makefile
|   |-- main.c
|   `-- implementation/
|       |-- implementation.c
|       `-- module.mk
`-- test/
    |-- Makefile
    `-- implementation/
        |-- implementation.test
        `-- module.mk

The source files are placed in modules under the src/ directory and is
organised in modules. The template contains one module:
'implementation'. Every module must contain a module.mk file that states
the source files in this directory and other necessary information such
as libraries required to build the module. The test files are located in
the test/ directory. They follow the same module structure as in src/.
Header files are placed in the include/ folder. This folder are included
in all compilations. 'local' header files can be placed in the module
folder.

The module.mk file under the test/ directory also contains a DEP_OBJ
variable, indicating required object to be linked to the test file in
this module.

To build with debug flags, run `make debug`. To analyse memory leaks and
such, run `make valgrind` which will run the program through valgrind.

Object and depend files are placed all over the place. To clean up, run
`make clean`.